@@ -1,7 +1,7 @@
 package com.timehop.stickyheadersrecyclerview.util;
 
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 
 /**
  * OrientationProvider for ReyclerViews who use a LinearLayoutManager
