@@ -1,6 +1,6 @@
 package com.timehop.stickyheadersrecyclerview.util;
 
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.RecyclerView;
 
 /**
  * Interface for getting the orientation of a RecyclerView from its LayoutManager
